♻️ [refactor] #129 로그인한 유저 기반으로 리팩토링
